backend/requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
python-multipart
pdfplumber
python-docx
tqdm
python-dotenv
